% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{are_concurrent_r}
\alias{are_concurrent_r}
\title{Check budget concurrency of a basket}
\usage{
are_concurrent_r(basket, utility_f, budget, rho = parent.frame())
}
\arguments{
\item{basket}{A numeric vector representing the quantities of goods in the basket.}

\item{utility_f}{An R function representing the utility function.}

\item{budget}{A numeric value specifying the maximum allowable utility limit.}

\item{rho}{The evaluation environment.}
}
\value{
A logical value (0 or 1).
}
\description{
This function checks whether the utility level of given basket is within given utility budget limit.
}
