% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{are_indifferent_r}
\alias{are_indifferent_r}
\title{Checking indifference of commodity baskets}
\usage{
are_indifferent_r(basket_A, basket_B, utility_f, rho = parent.frame())
}
\arguments{
\item{basket_A}{A numeric vector representing the quantities of goods in basket A.}

\item{basket_B}{A numeric vector representing the quantities of goods in basket B.}

\item{utility_f}{An R function that accepts a vector of goods and returns the utility level.}

\item{rho}{The evaluation environment for the utility function.}
}
\value{
A logical value (0 or 1).
}
\description{
This function verifies whether two consumption baskets have the same level of utility to the consumer.
}
