% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{classify_good_r}
\alias{classify_good_r}
\title{Economic classification of a good}
\usage{
classify_good_r(
  good,
  prices,
  income_or_non_wage,
  wage,
  max_time,
  utility_f,
  rho = 1e-04,
  model_type = 1
)
}
\arguments{
\item{good}{Index of the good being classified.}

\item{prices}{A numeric vector of commodity prices.}

\item{income_or_non_wage}{Nominal income or non-wage income.}

\item{wage}{Given period wage rate.}

\item{max_time}{Maximum time endowment.}

\item{utility_f}{A utility function implemented in R.}

\item{rho}{Numerical differentiation step size.}

\item{model_type}{Model type: 1 for classical model, 2 for labor market model.}
}
\value{
A character string (the economic classification of the good).
}
\description{
Automatically classifies the i-th good based on its income and price elasticity of demand.
Possible categories are:
* normal goods (an increase in consumer income leads to an increase in demand for the i-th good). They divide into:
     * basic (essential) - income elasticity in [0,1]
     * luxury - income elasticity greater than 1
* inferior goods (an increase in consumer income causes a decrease in demand for the i-th good) - income elasticity lesser than 0
There are also specific kind of goods, that behave unexpectedly:
* Giffen's good - type of inferior good, as the price of this good rises, so does the demand for it.
* Veblen's good - type of luxury good, as the price of this good rises, so does the demand for it. Indicator of a 'snob effect'.
}
