% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{first_Gossen_law_r}
\alias{first_Gossen_law_r}
\title{Verification of Gossen's first law}
\usage{
first_Gossen_law_r(utility_f, variable, rho = 1e-04)
}
\arguments{
\item{utility_f}{A utility function implemented in R.}

\item{variable}{A single-element vector (point x0) for derivative evaluation.}

\item{rho}{Numerical differentiation step size.}
}
\value{
A logical value (0 or 1).
}
\description{
Verifies a ceteris paribus - an increase in the consumption of the i-th good or service, while the consumption of other goods
and services remains constant, leads to increasingly smaller increases in total utility, 
accompanied by decreases in the marginal utility of the i-th good
}
