% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{Hicks_price_flexibility_of_demand_r}
\alias{Hicks_price_flexibility_of_demand_r}
\title{Hicksian Price Elasticity of Demand}
\usage{
Hicks_price_flexibility_of_demand_r(
  prices,
  target_utility,
  wage,
  max_time,
  utility_f,
  rho = 0.001,
  model_type = 1
)
}
\arguments{
\item{prices}{A numeric vector of commodity prices.}

\item{target_utility}{The target utility level.}

\item{wage}{The wage rate.}

\item{max_time}{Maximum time.}

\item{utility_f}{The utility function.}

\item{rho}{GSL machine numerical differentiation step size.}

\item{model_type}{Model type (1 or 2).}
}
\value{
A vector of compensated price elasticities of demand.
}
\description{
Calculates the vector of compensated (Hicksian) price elasticities of demand.
Properties:
* is continuous and increasing
* we pay nothing for a zero basket
*  if everything becomes x times more expensive, we will pay x times more.
}
