% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{marginal_utility_i_by_j_r}
\alias{marginal_utility_i_by_j_r}
\title{Marginal rate of substitution (MRS)}
\usage{
marginal_utility_i_by_j_r(utility_f, basket, i, j, rho = 1e-04)
}
\arguments{
\item{utility_f}{A utility function implemented in R.}

\item{basket}{A vector of commodity quantities.}

\item{i}{Index of the first good.}

\item{j}{Index of the second good.}

\item{rho}{Numerical differentiation step size.}
}
\value{
A numeric value specifying the MRS.
}
\description{
Calculates the Marginal Rate of Substitution of good i for good j.
It indicates by how much (approximately) the quantity of the jth good in basket should be increased when the
quantity of the ith good is reduced by one unit, so that the utility of the basket remains unchanged.
}
