% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{marginal_utility_r}
\alias{marginal_utility_r}
\title{Marginal utility}
\usage{
marginal_utility_r(utility_f, basket, rho = 1e-04)
}
\arguments{
\item{utility_f}{A utility function implemented in R.}

\item{basket}{A vector of commodity quantities.}

\item{rho}{Numerical differentiation step size (h_val).}
}
\value{
A numeric vector containing MU values for each good.
}
\description{
Calculates the vector of marginal utilities for all goods in the basket.
The marginal utility of the i-th good indicates by how much the utility of the basket will change if
the quantity of the i-th good increases by one unit, while the quantities of the other goods remain unchanged.
If the consumption of the i-th good or service increases while the consumption of the other goods
and services remains constant, then the total utility also increases.
}
