% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{maximize_utility_universal_r}
\alias{maximize_utility_universal_r}
\title{Universal consumer utility maximization}
\usage{
maximize_utility_universal_r(
  prices,
  income_or_non_wage,
  wage,
  max_time,
  utility_f,
  rho = parent.frame(),
  model_type = 1
)
}
\arguments{
\item{prices}{A numeric vector of commodity prices.}

\item{income_or_non_wage}{Nominal income or non-wage income.}

\item{wage}{Given period wage rate.}

\item{max_time}{Maximum time endowment.}

\item{utility_f}{A utility function implemented in R.}

\item{rho}{The evaluation environment.}

\item{model_type}{Model type: 1 for classical model, 2 for labor market model.}
}
\value{
A vector of commodity quantities representing the optimal demand basket.
}
\description{
Solves the consumer utility optimization problem using the Nelder-Mead algorithm.
Calculated consumer demand is being called Marshall's / Walras demand function
}
