% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{price_flexibility_of_demand_i_by_j_r}
\alias{price_flexibility_of_demand_i_by_j_r}
\title{Cross-Price elasticity of demand}
\usage{
price_flexibility_of_demand_i_by_j_r(
  prices,
  i_good,
  j_price,
  income_or_non_wage,
  wage,
  max_time,
  utility_f,
  rho = 1e-04,
  model_type = 1
)
}
\arguments{
\item{prices}{A numeric vector of commodity prices.}

\item{i_good}{Index of good i.}

\item{j_price}{Index of the price of good j.}

\item{income_or_non_wage}{Nominal income or non-wage income.}

\item{wage}{Given period wage rate.}

\item{max_time}{Maximum time endowment.}

\item{utility_f}{A utility function implemented in R.}

\item{rho}{Numerical differentiation step size.}

\item{model_type}{Model type: 1 for classical model, 2 for labor market model.}
}
\value{
A numeric value representing the cross-price elasticity.
}
\description{
Calculates the elasticity of demand for good i with respect to a price change of good j.
How the change of good j price changes how much good i a customer buys.
}
