% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{price_flexibility_of_demand_r}
\alias{price_flexibility_of_demand_r}
\title{Price elasticity of demand}
\usage{
price_flexibility_of_demand_r(
  prices,
  income_or_non_wage,
  wage,
  max_time,
  utility_f,
  rho = 1e-04,
  model_type = 1
)
}
\arguments{
\item{prices}{A numeric vector of commodity prices.}

\item{income_or_non_wage}{Nominal income or non-wage income.}

\item{wage}{Given period wage rate.}

\item{max_time}{Maximum time endowment.}

\item{utility_f}{A utility function implemented in R.}

\item{rho}{Numerical differentiation step size.}

\item{model_type}{Model type: 1 for classical model, 2 for labor market model.}
}
\value{
A vector of price elasticities.
}
\description{
Calculates the own-price elasticity of demand for individual goods in the basket.
Measures what percentage the demand for good i will change when its price changes by one percent.
}
