% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{price_index_r}
\alias{price_index_r}
\title{Price indexes (Laspeyres and Paasche)}
\usage{
price_index_r(base_basket, base_price, today_basket, today_price)
}
\arguments{
\item{base_basket}{A vector of quantities in the base period (q_0).}

\item{base_price}{A vector of prices in the base period (p_0).}

\item{today_basket}{A vector of quantities in the current period (q_t).}

\item{today_price}{A vector of prices in the current period (p_t).}
}
\value{
A numeric vector of length 2: c(Laspeyres, Paasche).
}
\description{
Calculates aggregate price change indices: the Laspeyres price index and the Paasche price index, that
indicate the percentage change in the cost of living. A price index value greater than 1 indicates an increase in 
the cost of living, while a value less than 1 indicates that the cost of living has fallen. The Laspeyres index measures
the cost of living based on a basket from the previous period, while the Paasche index measures it based on a basket from the current period.
}
