% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{quantity_index_r}
\alias{quantity_index_r}
\title{Quantity indexes (Laspeyres and Paasche)}
\usage{
quantity_index_r(base_basket, base_price, today_basket, today_price)
}
\arguments{
\item{base_basket}{A vector of quantities in the base period (q_0).}

\item{base_price}{A vector of prices in the base period (p_0).}

\item{today_basket}{A vector of quantities in the current period (q_t).}

\item{today_price}{A vector of prices in the current period (p_t).}
}
\value{
A numeric vector of length 2: c(Laspeyres, Paasche).
}
\description{
Calculates aggregate quantity change indices: the Laspeyres quantity index and the Paasche quantity index.
If index > 1, one can say that “average” consumption increased between time b and time t. It indicates an increase in consumption or an improvement in welfare.
If index < 1, one can say that “average” consumption decreased between periods b and t. It indicates a decline in welfare due to a decrease in consumption.
Laspeyres index < 1 indicates that the consumer prefers the basket from the previous period over the basket from the current period; Their position was better in the base period than in period t
Paasche index > 1 indicates that the consumer prefers the current period basket to the previous-period basket; Their position in period t is better than it was in the base period.
}
