% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{relationship_between_goods_r}
\alias{relationship_between_goods_r}
\title{Analysis of economic relationships between goods}
\usage{
relationship_between_goods_r(
  prices,
  i_good,
  j_good,
  i_price,
  j_price,
  income_or_non_wage,
  wage,
  max_time,
  utility_f,
  rho = 1e-04,
  model_type = 1
)
}
\arguments{
\item{prices}{A numeric vector of commodity prices.}

\item{i_good}{Index of the first good.}

\item{j_good}{Index of the second good.}

\item{i_price}{Index of the price of the first good.}

\item{j_price}{Index of the price of the second good.}

\item{income_or_non_wage}{Nominal income or non-wage income.}

\item{wage}{Given period wage rate.}

\item{max_time}{Maximum time endowment.}

\item{utility_f}{A utility function implemented in R.}

\item{rho}{Numerical differentiation step size.}

\item{model_type}{Model type: 1 for classical model, 2 for labor market model.}
}
\value{
A character string indicating the relationship (e.g., substitutes).
}
\description{
Determines the economic relationship between a pair of goods using cross-price elasticities.
Possible categories are:
* substitues - up to a factor of scale are interchangeable, they satisfy the same needs
* complementary - they are always consumed in equal proportions, any excess is wasted
* neutral - a change in the price of one of them does not affect the demand for the other 
* none of the above
}
