% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{reservation_wage_r}
\alias{reservation_wage_r}
\title{Reservation wage}
\usage{
reservation_wage_r(prices, income_non_wage, max_time, utility_f, rho = 1e-04)
}
\arguments{
\item{prices}{A vector of consumer commodity prices.}

\item{income_non_wage}{Non-wage income.}

\item{max_time}{Maximum available time.}

\item{utility_f}{The utility function.}

\item{rho}{Numerical differentiation step size.}
}
\value{
A numeric value representing the reservation wage.
}
\description{
Determines the reservation wage in the labor supply model.
It is the minimal wage above which a worker is willing to accept a job.
}
