% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{second_Gossen_law_r}
\alias{second_Gossen_law_r}
\title{Verification of Gossen's second law}
\usage{
second_Gossen_law_r(utility_f, basket, i, j, prices, rho = 1e-04)
}
\arguments{
\item{utility_f}{A utility function implemented in R.}

\item{basket}{The evaluated basket of goods.}

\item{i}{Index of good i.}

\item{j}{Index of good j.}

\item{prices}{A vector of market prices.}

\item{rho}{Numerical differentiation step size.}
}
\value{
A logical value (0 or 1).
}
\description{
Verifies the consumer equilibrium condition – in order to maximize utility, the consumer chooses a basket
of goods in which the ratios of the marginal utilities of these goods are equal to the ratio of their prices.
The higher the price of the substitute good , the higher the marginal rate of substitution (assuming that the price pj remains constant).
}
