% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{substitution_flexibility_i_by_j_r}
\alias{substitution_flexibility_i_by_j_r}
\title{Relative elasticity of utility substitution}
\usage{
substitution_flexibility_i_by_j_r(utility_f, basket, i, j, rho = 1e-04)
}
\arguments{
\item{utility_f}{A utility function implemented in R.}

\item{basket}{A vector of commodity quantities.}

\item{i}{Index of the first good.}

\item{j}{Index of the second good.}

\item{rho}{Numerical differentiation step size.}
}
\value{
A numeric value.
}
\description{
Calculates the ratio of utility elasticity of good i to good j.
Indicates the approximate percentage by which the quantity of the jth good in basket must be increased
while reducing the quantity of the ith good by one percent, so that the utility of that basket remains unchanged.
}
