% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{total_changes_index_r}
\alias{total_changes_index_r}
\title{Total value changes index}
\usage{
total_changes_index_r(base_basket, base_price, today_basket, today_price)
}
\arguments{
\item{base_basket}{A vector of quantities in the base period (q_0).}

\item{base_price}{A vector of prices in the base period (p_0).}

\item{today_basket}{A vector of quantities in the current period (q_t).}

\item{today_price}{A vector of prices in the current period (p_t).}
}
\value{
A numeric value representing the value index.
}
\description{
Calculates the ratio of total expenses in the current period to the base period.
If Paasche price index > index then the consumer prefers basket from period b to basket x from current period
If Laspeyres price index < index then the consumer prefers basket from current period to basket x from period b
}
