% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{usage_flexibility_r}
\alias{usage_flexibility_r}
\title{Elasticity of utility with respect to good consumption}
\usage{
usage_flexibility_r(utility_f, basket, i, rho = 1e-04)
}
\arguments{
\item{utility_f}{A utility function implemented in R.}

\item{basket}{A vector of commodity quantities.}

\item{i}{Index of the first good.}

\item{rho}{Numerical differentiation step size.}
}
\value{
A numeric value representing the elasticity.
}
\description{
Calculates the partial elasticity of the utility function with respect to the i-th good.
Indicates the approximate percentage change in the utility of basket if the quantity of the i-th good
increases by one percent, while the quantities of the other goods remain unchanged.
}
